+++ modules/rgbd/src/tsdf.cpp
@@ -5,8 +5,7 @@
 // This code is also subject to the license terms in the LICENSE_KinectFusion.md file found in this module's directory
 
 #include "precomp.hpp"
-//#include "tsdf.hpp"
-#include "tsdf_functions.hpp"
+#include "tsdf.hpp"
 #include "opencl_kernels_rgbd.hpp"
 
 namespace cv {
